:recycle: Refactoring code. Corrected error InvalidRequestError: One or more mappers failed to initialize - can't proceed with initialization of other mappers. Triggering mapper: 'Mapper[Booking(booking)]'. Original exception was: Mapper 'Mapper[Event(event)]' has no property 'bookings'. If this property was indicated from other mappers or configure events, ensure registry.configure() has been called. в модели booking and events
