+++ alembic/versions/b09e99c70f66_initial_migrations.py
@@ -1,0 +1,26 @@
+"""Initial migrations
+
+Revision ID: b09e99c70f66
+Revises: 
+Create Date: 2024-09-06 16:44:35.888714
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'b09e99c70f66'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
