--- alembic/versions/ec0bd3e38712_added_is_hidden_field_to_event_model.py
@@ -1,30 +1,0 @@
-"""Added is_hidden field to event model
-
-Revision ID: ec0bd3e38712
-Revises: 6a62417341a1
-Create Date: 2024-09-01 16:09:28.087839
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ec0bd3e38712'
-down_revision: Union[str, None] = '6a62417341a1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('event', sa.Column('is_hidden', sa.Boolean(), server_default=sa.text('0'), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('event', 'is_hidden')
-    # ### end Alembic commands ###
