+++ src/project/templates/register.html
@@ -20,13 +20,13 @@
                         <div class="mb-4">
                             <label for="email" class="block mb-2 text-left">Email:</label>
                             <input type="email"
-                                   class="w-3/4 p-2 rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
+                                   class="w-3/4 p-2 text-black rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
                                    id="email" name="email" required>
                         </div>
                         <div class="mb-4">
                             <label for="password" class="block mb-2 text-left">Password:</label>
                             <input type="password"
-                                   class="w-3/4 p-2 rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
+                                   class="w-3/4 p-2 text-black rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
                                    id="password"
                                    name="password" required>
                         </div>
@@ -35,7 +35,7 @@
 
                             <label for="confirm_password" class="block mb-2 text-left">Confirm Password:</label>
                             <input type="password"
-                                   class="w-3/4 p-2 rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
+                                   class="w-3/4 p-2 text-black rounded-md border-gray-600 focus:ring focus:ring-indigo-500"
                                    id="confirm_password" name="confirm_password" required>
                         </div>
